fix compiler warnings: enum-enum mixing, u8path deprecation, GCC 16 nonnull FP

+++ tools/mtmd/clip-graph.h
@@ -9,7 +9,9 @@
 #include <vector>
 #include <functional>
 
-#define DEFAULT_INTERPOLATION_MODE (GGML_SCALE_MODE_BILINEAR | GGML_SCALE_FLAG_ANTIALIAS)
+// flags live in a separate enum, so combine them as integers to avoid
+// -Wdeprecated-enum-enum-conversion
+#define DEFAULT_INTERPOLATION_MODE ((int) GGML_SCALE_MODE_BILINEAR | (int) GGML_SCALE_FLAG_ANTIALIAS)
 
 struct build_vit_opts {
     ggml_tensor * attn_mask = nullptr;
